backend/requirements.txt
# ---- Core ----
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ---- Settings ----
python-dotenv==1.0.1
pydantic-settings==2.2.1

# ---- Database ----
SQLAlchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# ---- Auth ----
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ---- Background / Cache ----
celery==5.3.6
redis==5.0.3

# ---- Utilities ----
python-multipart==0.0.9
email-validator==2.1.0.post1

# ---- Testing ----
pytest==8.0.2
httpx==0.27.0
